EOS-12728: Create separate CortxCli RPM     * Fix review notes     * fix build process     * rename cc_setup.py to cortx_cli_setup.py     * clean-up
